updated authorization

+++ app/controllers/sessions_controller.rb
@@ -1,9 +1,12 @@
 class SessionsController < ApplicationController
+  skip_before_action :authorized, only: :create
+
   def create
-    # byebug
     user = User.find_by(username: params[:username])
     if user&.authenticate(params[:password])
       session[:user_id] = user.id
+      #think below should be changed without the p ;)
+      p session[:user_id]
       render json: user, status: :ok
     else 
       render json: {errors: ["Invalid username or password"]}, status: :unauthorized
@@ -11,7 +14,9 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    session.delete(:user_id)
-    head :no_content
+    if current_user 
+      session.delete(:user_id)
+      head :no_content
+    end
   end
 end
